Rename crash handler files
