+++ interface/src/CrashHandler.h
@@ -2,8 +2,8 @@
 //  CrashHandler.h
 //  interface/src
 //
-//  Created by David Rowe on 24 Aug 2015.
-//  Copyright 2015 High Fidelity, Inc.
+//  Created by Clement Brisset on 01/19/18.
+//  Copyright 2018 High Fidelity, Inc.
 //
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
@@ -12,22 +12,21 @@
 #ifndef hifi_CrashHandler_h
 #define hifi_CrashHandler_h
 
-#include <QString>
+#include <string>
 
-class CrashHandler {
+#if HAS_CRASHPAD
 
-public:
-    static bool checkForResetSettings(bool wasLikelyCrash, bool suppressPrompt = false);
+bool startCrashHandler();
+void setCrashAnnotation(std::string name, std::string value);
 
-private:
-    enum Action {
-        DELETE_INTERFACE_INI,
-        RETAIN_IMPORTANT_INFO,
-        DO_NOTHING
-    };
+#elif HAS_BREAKPAD
 
-    static Action promptUserForAction(bool showCrashMessage);
-    static void handleCrash(Action action);
-};
+bool startCrashHandler();
+void setCrashAnnotation(std::string name, std::string value);
 
-#endif // hifi_CrashHandler_h
+#else
+
+bool startCrashHandler();
+void setCrashAnnotation(std::string name, std::string value);
+
+#endif // hifi_Crashpad_h
